add smart abstract class
